==> data_prep/pipeline/01_gen_cleaned_rows.ipynb <==
import pandas as pd, numpy as np

import pickle

# LSFF: Output data counts under different binning scenarios

## read in data
path = '/ihme/homes/beatrixh/vivarium_research_lsff/data_prep/outputs/lsff_data_stage0_3_24_2021.3.csv'
df = pd.read_csv(path)

df.head()

## for convenience
check_cols = ['location_id','location_name','vehicle','value_description','nutrient','value_mean','value_025_percentile',
       'value_975_percentile','notes']

df.columns

df.loc[(df.location_name=="Vietnam") & (df.vehicle=="wheat flour"),['location_name','vehicle','nutrient','value_description','nutrient','value_mean','source_year', 
                                         'sub_population','source_citation', 'source_link', 'source_type']]

df.loc[df.location_name=="Burkina Faso",['location_name','vehicle','nutrient','value_description','nutrient','value_mean','source_year', 
                                         'sub_population','source_citation', 'source_link', 'source_type']]

for i in df.loc[df.location_name=="Burkina Faso",['location_name','vehicle','nutrient','value_description','nutrient','value_mean','source_additional', 'source_citation', 'source_link', 'source_type',
       'source_year', 'sub_population']].source_link.unique():
    print(i)

## check that we have nutrients iff expected

df[df.nutrient=='na'].value_description.unique()

df[df.nutrient!='na'].value_description.unique()

df[df.value_description=="percent of population eating adequately fortified vehicle"]

# relabel vals

#https://journals.sagepub.com/doi/pdf/10.1177/15648265120334S307
# cote divoire,
# "During the last year of the project, 206,410 MT of
# vegetable oil was marketed in Côte d’Ivoire. The fortified
# vegetable oil was covering 89% of the market."

val_descrip_map = {'percent of population eating fortified vehicle': 'percent of population eating fortified vehicle',
                   'percent of population eating adequately fortified vehicle': 'percent of population eating fortified vehicle',
 'percent of vehicle that is industrially produced': 'percent of vehicle that is industrially produced',
 'percent of vehicle that is fortified': 'percent of vehicle that is fortified',
 'percent of population eating industrially produced vehicle': 'percent of population eating industrially produced vehicle',
 'percent of marketshare of fortified products': 'percent of market covered by fortified product',
 'percent of population eating vehicle': 'percent of population eating vehicle'}

df.value_description = df.value_description.map(val_descrip_map)

df.value_description.unique()

df.head()

df.columns

def output_frame(cols= ['location_name','vehicle','nutrient','value_description']):
    """
    INPUT: desired columns to bin lsff data on
    ---
    OUTPUT: all legal value combinations using input columns
    ---
    eg: (iodine/salt) or (china/maize flour) are illegal combinations
    """
    
    # load legal pairs
    data_prep_dir = '/ihme/homes/beatrixh/vivarium_research_lsff/data_prep/inputs/'
    with open(data_prep_dir + 'lsff_vehicle_nutrient_pairs.pickle', 'rb') as handle:
        vehicle_nutrient_map = pickle.load(handle)

    with open(data_prep_dir + 'lsff_country_vehicle_pairs.pickle', 'rb') as handle:
        country_vehicle_map = pickle.load(handle)
        
    countries = [i for i in country_vehicle_map.keys() if type(i)==str]
    
    # value_descrips
    nutrient_relevant_vds = ['percent of population eating fortified vehicle',
           'percent of vehicle that is fortified',
           'percent of market covered by fortified product']
    nutrient_irrelevant_vds = ['percent of vehicle that is industrially produced',
       'percent of population eating industrially produced vehicle',
       'percent of population eating vehicle']

    estimate_status = [0,1]
    
    # all legal loc-vehicle-nutrient-val_ds for val_ds that have nutrient
    data_counts_a = pd.DataFrame([(i,j,k,l) for i in countries
                                  for j in country_vehicle_map[i]
                                  for k in vehicle_nutrient_map[j]
                                  for l in nutrient_relevant_vds],
                                 columns = ['location_name','vehicle','nutrient','value_description'])
    # all legal loc-vehicle-nutrient-val_ds for val_ds with nutrient='na'
    data_counts_b = pd.DataFrame([(i,j,'na',k) for i in countries
                                  for j in country_vehicle_map[i]
                                  for k in nutrient_irrelevant_vds],
                                 columns = ['location_name','vehicle','nutrient','value_description']) 
    
    data_counts = data_counts_a.append(data_counts_b).sort_values(by=['location_name','vehicle','value_description','nutrient'])
    
    return data_counts[cols].drop_duplicates().reset_index(drop=True)

def add_summary_measures_helper(frame, data, n_points = True, mean_mean = True, mean_values = True):
    """
    INPUT:
    - df with extracted data containing a 'value_mean' column, and a set of id_cols
    - a frame whose cols == id_cols
    - desired summary measures for each unique combination of id_cols
    """
    
    # merge extracted data onto frame containing desired bins
    merge_cols = frame.columns.tolist()
    data_counts = frame.merge(data[merge_cols + ['value_mean']], on = merge_cols, how = 'left')
    output = frame.copy()
    
    # calculate summary measures
    if n_points:
        data_counts_n = data_counts.groupby(merge_cols).count().rename(columns={'value_mean':'n_data_points'})
        output = output.merge(data_counts_n, on = merge_cols, how = 'left')
    if mean_mean:
        data_counts_mean = data_counts.groupby(merge_cols).mean().rename(columns={'value_mean':'mean_mean'})
        output = output.merge(data_counts_mean, on = merge_cols, how = 'left')
    if mean_values:
        data_counts_detail = data_counts.groupby(merge_cols).aggregate(lambda x : x.tolist()).rename(columns={'value_mean':'value_means'})
        output = output.merge(data_counts_detail, on = merge_cols, how = 'left')
        
    return output

def add_summary_measures_handling_estimates(frame, data, n_points = True, mean_mean = True, mean_values = True):
    """
    INPUT:
    - df with extracted data containing a 'value_mean' column, and a set of id_cols
    - a frame whose cols == id_cols
    - desired summary measures for each unique combination of id_cols
    -----
    @returns (df_A, df_B):
    df_A:
    - a row for each unique combination of id_cols
    - for each n_points, mean_mean, mean_values set to True, a col of corresponding vals
    - a column is_assumption, flagging rows for which we didn't have data, but made an assumption
    df_B:
    - all rows of dropped data.
    
    @effects
    - for any value where we have both data and an assumption, drops the assumption (and
    return as part of df_B)
    """
    
    #helpers
    merge_cols = frame.columns.tolist()
    return_cols = merge_cols
    
    #split out into extracted data vs assumptions
    extracted = add_summary_measures_helper(frame, df[df.is_estimate!=1], True, mean_mean, mean_values)
    assm = add_summary_measures_helper(frame, df[df.is_estimate==1], True, mean_mean, mean_values)
    
    #wide on is_estimated status
    data = extracted.merge(assm, on = merge_cols, suffixes = ['','_assm'], how = 'outer')
    assert(len(data)==len(extracted)), "merge issue"
    assert(len(data)==len(assm)), "merge issue"
    
    #rows for which we have both extracted and assumed "data"
    both = data[(data.n_data_points > 0) & (data.n_data_points_assm > 0)]
    if len(both)>0:
        print("(!) Make sure you are okay throwing out datapoint estimates for all rows for which we also have extracted data")
    
    #we will throw out the assumed data, and return this df to double check
    data.loc[(data.n_data_points > 0) & (data.n_data_points_assm > 0),'n_data_points_assum'] = 0
    
    
    #for all assumed data where we dont have extracted data, keep and mark as such
    data.loc[(data.n_data_points == 0) & (data.n_data_points_assm > 0),'is_assumption'] = 1
    data.loc[(data.is_assumption==1),'n_data_points'] = data.n_data_points_assm
    if n_points:
        return_cols = return_cols + ['n_data_points']
    if mean_mean:
        data.loc[(data.is_assumption==1),'mean_mean'] = data.mean_mean_assm
        return_cols = return_cols + ['mean_mean']
    if mean_values:
        data.loc[(data.is_assumption==1),'value_means'] = data.value_means_assm
        return_cols = return_cols + ['value_means']
    
    data.loc[data.is_assumption!=1,'is_assumption'] = 0
    data.is_estimate = data.is_assumption.astype(int)
        
    return data[return_cols + ['is_assumption']], both

## count data coverage for ALL county - vehicle - val_des - nutrient combinations

data_counts = output_frame(cols = ['location_name','vehicle','nutrient','value_description'])

data_counts_full, check = add_summary_measures_handling_estimates(frame = data_counts, data = df, n_points = True, mean_mean = True, mean_values = True)

data_counts_full

#going to drop these vals where we assumed zero
check

data_counts_full_path = '/ihme/homes/beatrixh/vivarium_research_lsff/data_prep/outputs/data_counts_full_3_24_2021.3.csv'
data_counts_full.to_csv(data_counts_full_path, index = False)

## save each row of data (not collapsed with summary stats)

data_rows = df[['location_name','vehicle','nutrient','value_description','value_mean','value_025_percentile','value_975_percentile','is_estimate','sub_population','source_year']]

#drop the row of assumed data
data_rows = data_rows[~((data_rows.location_name=="Ethiopia") &
         (data_rows.vehicle=="salt") &
         (data_rows.nutrient=="folic acid") &
         (data_rows.is_estimate==1.0))]

# clear illegal combos
data_rows = data_rows.merge(data_counts, on = ['location_name','vehicle','nutrient','value_description'], how = 'inner')

data_rows = data_rows.sort_values(by=['location_name','vehicle','value_description','nutrient'])

data_rows_path = '/ihme/homes/beatrixh/vivarium_research_lsff/data_prep/outputs/population_coverage_data_3_24_2021.3.csv'
data_rows.to_csv(data_rows_path, index = False)

## count data coverage for target county - vehicle - val_des - nutrient combinations

target_value_descriptions = ['percent of population eating fortified vehicle',
       'percent of population eating industrially produced vehicle',
       'percent of population eating vehicle']

target_data_counts = data_counts_full[data_counts_full.value_description.isin(target_value_descriptions)]

target_data_counts

target_data_counts_path = '/ihme/homes/beatrixh/vivarium_research_lsff/data_prep/outputs/data_counts_ideal_bins_3_24_2021.3.csv'
target_data_counts.to_csv(target_data_counts_path, index = False)

## count data coverage for ALL county - vehicle - val_des combos, NOT nutrients

data_counts_nutrient_na = output_frame(cols = ['location_name','vehicle','value_description'])

data_counts_nutrient_na, check_b = add_summary_measures_handling_estimates(data_counts_nutrient_na, df, n_points = True, mean_mean = True, mean_values = True)

data_counts_nutrient_na

check_b

data_counts_nutrient_na_path = '/ihme/homes/beatrixh/vivarium_research_lsff/data_prep/outputs/data_counts_nutrient_na_3_24_2021.3.csv'
data_counts_nutrient_na.to_csv(data_counts_nutrient_na_path, index = False)
