==> data_prep/pipeline/03a_combine_cleaned_rows_with_data_guesses.ipynb <==
import pandas as pd, numpy as np, matplotlib.pyplot as plt

# LSFF: combine 01_gen_cleaned_rows.output() with 02b_first_pass_fill_data_gaps.output()


# note that this is different from the output of 02b_first_pass_fill_data_gaps.output(), which includes only one row per
# value we're looking for.  01_gen_cleaned_rows.output(), on the other hand, includes all cleaned data we have (even 
# if it's duplicative for a given value we're looking for)

data_path = '/ihme/homes/beatrixh/vivarium_research_lsff/data_prep/outputs/population_coverage_data_3_24_2021.3.csv'
df = pd.read_csv(data_path)

df.head()

estimates_path = '/ihme/homes/beatrixh/vivarium_research_lsff/data_prep/outputs/lsff_data_estimated_03_26_2021.4.csv'
est = pd.read_csv(estimates_path)

est_val_map = {'A': 'percent of population eating vehicle',
 'B': 'percent of population eating industrially produced vehicle',
 'C': 'percent of population eating fortified vehicle',
 'D': 'percent of vehicle that is industrially produced',
 'E': 'percent of vehicle that is fortified',
 'F': 'percent of market covered by fortified product'}

est.head()

est.dropna()

est.standard.unique()

est = est[(est.B_estimate==1) | (est.C_estimate==1)]

#keep estimates only; column A is all real data
est = pd.melt(est, id_vars = ['location_name','vehicle','nutrient','standard'], value_vars = ['B','C'],
       var_name = 'value_description', value_name = 'value_mean')

est.value_description = est.value_description.map(est_val_map)

est['is_estimate'] = 1

df = df.append(est, sort = False)

df.head()

save_path = '/ihme/homes/beatrixh/vivarium_research_lsff/data_prep/outputs/population_coverage_data_and_estimates_3_26_2021.4.csv'
df.to_csv(save_path, index = False)
